Move where we do metrics, since we can't get the path properly via the middleware
