+++ MorphicServer/Startup.cs
@@ -61,7 +61,6 @@ namespace MorphicServer
             app.UseRouting();
             app.UseSerilogRequestLogging();
             //app.UseHttpMetrics(); // doesn't work. Probably because we have our own mapping, and something is missing
-            app.UseRequestMiddleware();
             app.UseEndpoints(Endpoint.All);
             app.UseEndpoints(endpoints =>
             {
